APP-25: move router to new file and guard protected routes
